fix: support bfx terminal trades
fixes an issue where a stream filter was applied to trades, even
though just unwanted candles need to get filtered out
