--- test/todo.js
@@ -1,1 +1,0 @@
-/* eslint-env mocha */
